+++ sample/sample22-closure-table/app.ts
@@ -1,5 +1,5 @@
 import "reflect-metadata";
-import {createConnection, CreateConnectionOptions} from "../../src/index";
+import {createConnection, CreateConnectionOptions} from "../../src/backend";
 import {Category} from "./entity/Category";
 
 const options: CreateConnectionOptions = {
