+++ typings.json
@@ -8,7 +8,6 @@
     "sinon": "registry:dt/sinon#1.16.0+20160517064723"
   },
   "globalDependencies": {
-    "es6-shim": "registry:dt/es6-shim#0.31.2+20160317120654",
     "node": "registry:dt/node#6.0.0+20160524002506"
   },
   "dependencies": {
